export classes from Foo object
